requirements.txt
openai>=1.0.0
diffusers>=0.20.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
requests>=2.31.0
numpy>=1.24.0
accelerate>=0.20.0
google-genai>=0.1.0
pillow>=10.0.0
google-cloud-aiplatform==1.46.0
